Finally have the homepage working
After hours of struggling, and trying to learn Angular reading and watching videos, I figured it out I think. I had to make a folder called shared and a file called bootstrap module so I can import the modules for ngx bootstrap then I had to import SharedBootstrapModule in AppModule.

In order to use the carousel on the home page I had to install ngX bootstrap  (I used the terminal). honestly I think the regular worked I just had the wrong slashes and the path was too specific. but it shows.

the nav bar I am using for the home page will change. I was trying to use this cool template but I am struggling with the css portion .

I will work on the about page sunday afternoon.